Cloning and serialization improvements (#21)
======================================

The main purpose of this commit is to increase
the correctness when cloning and/or pickling
instances of Problem, Solution, SolutionBatch,
ObjectArray, and ImmutableContainer.
When cloning and/or pickling, two new behaviors
are introduced:
(i) PyTorch tensors are cut off from their
original storages (to avoid small slices from
carrying their original storages with
themselves); and
(ii) self references within these mentioned
objects are specially handled, and therefore,
many unwanted RecursionError scenarios which
could previously happen are now avoided.

New function: `deep_clone(...)`
===============================

This new function creates a deep copy of a
Python container (which can contain other
containers or PyTorch tensors or numpy arrays).
While traversing through the container, any
PyTorch tensors will be copied using the
`.clone()` method, which means, in the newly
created clone, the PyTorch tensors will be
independent copies without extra storage.

New base class: `Clonable`
==========================

Any subclass of `Clonable` is expected to
override the method `_get_cloned_state(...)`.
This method is expected to return the state
of the object in a dictionary.
The `Clonable` base class provides definitions
for the methods `__copy__(...)` and
`__deepcopy__(...)` where the dictionary made
by `_get_cloned_state(...)` is used to create
the clone.

New base class: `Serializable`
==============================

This is a subclass of Clonable, which is meant
to be used as a base class. In addition to the
method definitions made by `Clonable` (
`__copy__` and `__deepcopy__`), this base class
also defines `__getstate__` so that the method
`_get_cloned_state(...)` also affects how the
inheriting class is pickled.

Classes inheriting from `Serializable`
======================================

The following classes now inherit from
`Serializable`:

 - Problem
 - Solution
 - SolutionBatch
 - Distribution

Each of these mentioned classes have their own
definitions for `_get_cloned_state(...)`, and
use `deep_clone(...)` for preparing their state
dictionaries. Therefore, at the moment of
cloning and pickling, the stored tensors are
subject to the `.clone()` method and they are
cut off from their original storages.
With this feature, when a relatively small
slice of a large SolutionBatch is cloned or
pickled, the serialized slice will not
contain the entire storage of the original
batch.

New base class: `ReadOnlyClonable`
==================================

This is like `Clonable`, but requires the
inheriting class to also define a method
called `_get_mutable_clone()` so that the
object knows how its mutable counterpart
can be constructed. ImmutableContainer
objects are now ReadOnlyClonable.