fix: the signature of `_do_cross_over(...)`
The `CrossOver` base class has a method
`_do_cross_over(...)` that is meant to be
overriden by the inheriting classes.
Although the documentation and the initial
signature of `_do_cross_over(...)` claims
that the expected arguments are
`batch` (the original SolutionBatch),
`parents1` (decision values of the first
half of the selected parents)
and `parents2` (decision values of the
second half of the selected parents),
what the class `CrossOver` actually does
is to call it only with the arguments
`parents1` and `parents2` (this can be
confirmed by looking at the `_do(...)`
method of `CrossOver`). All our `CrossOver`
subclasses also define their implementations
of `_do_cross_over(...)` with the arguments
`parents1` and `parents2` only.
Therefore, the initial signature of
`_do_cross_over(...)` and its docstring
are misleading. This commit aims to fix
those misleading elements.