chore: Fix infinite live reloading of the docs
Previously, the docs would reload infinitely because there was a hook
that automatically generated markdown files out of jupyter notebooks.

The problem was that the generated markdown files were stored directly
into docs folder, which triggered mkdocs to rebuild again.

This patch changes the target directory of the generated files
and adds a new mkdocs hook (`on_files`) that adds all the files
during the build process. This way the autogenerated files
are not triggering the reload.