fix: Solution.set_evals with eval data
When working on a problem with a non-zero
`eval_data_length`, the `set_evals(...)`
method of the `Solution` object was failing
when the user passed the extra evaluation
data separately as a 1-dimensional tensor.
This was inconsistent with the user interface
documented within the docstrings of
`Solution.set_evals(...)`.

This commit fixes this inconsistency.

Co-authored-by: Pawel Liskowski <pawel@nnaisense.com>