fix: Interface of `make_I` (#62)
* Fix the interface of `make_I`

The function `evotorch.tools.make_I` was
accepting a single integer as the size
indicator of the identity matrix. However,
The method `Problem.make_I` was working only
when the size was given as a single-element
tuple. This commit makes their interfaces
compatible by making isure that both will
work with a single integer or with a size
tuple containing a single integer.
Tests are updated to verify the correctness
of the revised interfaces.

* Update initialisation of identity matrices in CMA-ES and XNES to use new API

* Have CMA-ES refer to internal problem instance

---------

Co-authored-by: Timothy Atkinson <dr.timothy.atkinson@gmail.com>