fix: get_minibatch() of SupervisedNE
This commit fixes/improves SupervisedNE in two
ways:

1)
The `get_minibatch()` method of `SupervisedNE`
no longer uses recursion to handle the end of
the data loader's minibatches. The new
implementation catches only `StopIteration`
and non-recursively restarts the iterator of the
data loader. Any other type of error is now
deliberately unhandled to avoid unwanted
infinite recursion and to allow the user to see
the details of the error.

2)
The example notebook `Training_MNIST30K.ipynb`
demonstrating the usage of `SupervisedNE` is
updated so that its hyperparameter configuration
follows what is reported in the technical report
of EvoTorch. This way, results reported in the
report can be reproduced.

Co-authored-by: Timothy Atkinson <timothy@nnaisense.com>
Co-authored-by: Nihat Engin Toklu <engin@nnaisense.com>