Make EvoTorch future-proof
Some of the dependencies of EvoTorch became
deprecated over time. This commit aims to
replace those deprecated requirements with their
up-to-date counterparts.
In more details, the following changes are
introduced:

- If available, use `torch.func` instead of
  `functorch`
- For getting the storage address of a tensor,
  use `x.untyped_storage().data_ptr()`
  instead of `x.storage().data_ptr()`,
  if available
- Use `gymnasium` instead of `gym`
- Introduce an EvoTorch-specific adapter between
  `gymnasium` and `brax`, instead of relying on
  the deprecated `gym` wrapper of `brax`

Status: after the changes introduced by this
commit, the unit tests pass successfully.
Apart from the unit tests, test runs are still
needed to ensure the correctness of these
changes.