Add ability to work with brax.v1 tasks
Brax v0.9.0 has a new set of implementations
for its tasks. The old ones (which were the
default until brax v0.1.2) are now within the
namespace `brax.v1`.
This commit extends `VecGymNE` with the ability
to instantiate tasks from the namespace
`brax.v1`, so that it will be possible to
reproduce older results.