Update the pybullet humanoid example
The pybullet humanoid example within the
directory `examples/scripts/` is updated so that
it works correctly with the new `GymNE`
(which now uses the `gymnasium` library).

At the moment of writing this, PyBullet
environments (within the module `pybullet_envs`)
do not have direct interfaces for `gymnasium`,
and they fail with the latest version of the
classical `gym` (i.e. gym v0.26).
Therefore, for the pybullet humanoid example to
work, one needs to install the classical `gym`
like this: `pip install 'gym<0.26'`
(while ensuring that `gymnasium` is also
installed).