fix: Wrong defaults of SteadyStateGA (#87)
While initializing `SteadyStateGA`, if the
`operators` argument was not given (or was given
as an empty list) and if `re_evaluate` was given
as True (which is the default), `SteadyStateGA`
was making the wrong default assumption that the
parent solutions do not have to be evaluated
before the children. This wrong assumption
caused the tournament selection phases of the
cross-over operators to fail, as they encountered
unevaluated parents.

This commit fixes this wrong assumption.
When `operators` is empty (which means that the
operators are to be added later via `use(...)`
method), and `re_evaluate` is True, then
`SteadyStateGA` will by default assume that the
parents need to be evaluated first.