feat: Make EvoTorch future-proof (#77)
* Make EvoTorch future-proof

Some of the dependencies of EvoTorch became
deprecated over time. This commit aims to
replace those deprecated requirements with their
up-to-date counterparts.
In more details, the following changes are
introduced:

- If available, use `torch.func` instead of
  `functorch`
- For getting the storage address of a tensor,
  use `x.untyped_storage().data_ptr()`
  instead of `x.storage().data_ptr()`,
  if available
- Use `gymnasium` instead of `gym`
- Introduce an EvoTorch-specific adapter between
  `gymnasium` and `brax`, instead of relying on
  the deprecated `gym` wrapper of `brax`

Status: after the changes introduced by this
commit, the unit tests pass successfully.
Apart from the unit tests, test runs are still
needed to ensure the correctness of these
changes.

* Update example reinforcement learning notebooks

This commit updates the example notebooks
demonstrating reinforcement learning and model
predictive control, so that they adopt the API
of the `gymnasium` library.

* Add ability to work with brax.v1 tasks

Brax v0.9.0 has a new set of implementations
for its tasks. The old ones (which were the
default until brax v0.1.2) are now within the
namespace `brax.v1`.
This commit extends `VecGymNE` with the ability
to instantiate tasks from the namespace
`brax.v1`, so that it will be possible to
reproduce older results.

* Update the rendering section of the brax example

* Fix the markdown sections in the MPC example

* Update the pybullet humanoid example

The pybullet humanoid example within the
directory `examples/scripts/` is updated so that
it works correctly with the new `GymNE`
(which now uses the `gymnasium` library).

At the moment of writing this, PyBullet
environments (within the module `pybullet_envs`)
do not have direct interfaces for `gymnasium`,
and they fail with the latest version of the
classical `gym` (i.e. gym v0.26).
Therefore, for the pybullet humanoid example to
work, one needs to install the classical `gym`
like this: `pip install 'gym<0.26'`
(while ensuring that `gymnasium` is also
installed).

* Fix: brax action boundaries are now correctly reported

* Use the new API of neptune

* Apply suggestions from code review

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: Vojtech Micka <vojtech@nnaisense.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>