docs: Inform the user when `device` is not set correctly (#90)
* Inform the user when `device` is not set correctly

In EvoTorch, the `device` of a problem must be
set as "cpu" if the problem is configured to have
multiple actors. With this commit, when this rule
is broken, a detailed error message is shown to
the user. The error message also informs the user
regarding what to do to fix the error.

---------

Co-authored-by: Vojtech Micka <vojtech@nnaisense.com>