feat: Add a functional API for optimization
This commit introduces an alternative API for
EvoTorch that conforms to the functional
programming paradigm. This functional API can
be used together with `torch.func.vmap`, and
therefore can be used for optimizing not just
a single population, but a batch of populations
simultaneously.

The main improvements are:

(i) Functional counterparts of cross entropy
method (`cem`) and policy gradients with
parameter-based exploration (`pgpe`) are
implemented. These algorithm implementations can
be used with `vmap`, or can be given batches of
starting points (`center_init` arguments) so that
they will generate batches of populations
centered around them.

(ii) Functional counterparts of gradient-based
optimizers `adam`, `clipup`, and `sgd` are
implemented. The interfaces of these optimizers
are similar to the interfaces of functional `cem`
and `pgpe`. Therefore, the user can switch back
and forth between evolutionary approach and
gradient-based approach for solving a problem,
with minimum amount of code change.

(iii) The decorator `@expects_ndim` is
introduced. This decorator is used for declaring
how many dimensions are expected per each
positional argument of the decorated function.
Upon receiving tensors whose number of dimensions
are more than expected, the decorated function
interprets those tensors as batched arguments,
applies `vmap` on itself, and performs its
operations across the batch dimensions.

(iv) The decorator `@rowwise` is introduced.
This decorator is used for declaring that a
function is implemented with the assumption that
its argument is a vector. If it receives a tensor
with 2 or more dimensions, the function applies
`vmap` on itself, and performs its operations
across multiple batch dimensions.