Add a notebook demonstrating how to evolve objects
This commit adds an example notebook which
demonstrates how one can declare and solve an
optimization problem that has its `dtype` set as
`object`, and therefore has a custom solution
structure.

While working on the example notebook, one bug,
and one genetic algorithm-related shortcoming
have been identified and fixed.

The bug was preventing one from instantiating a
new `ObjectArray` with the help of
`evotorch.tools.as_tensor(..., dtype=object)`.
This commit fixes this bug.

The shortcoming was as follows: when the user
defined a custom mutation operator using a
regular function, `GeneticAlgorithm` assumed
that the regular function would return a new
batch of solutions whose length is equal to the
number of solutions it received. This was an
unnecessary restriction, and therefore, got
fixed with this commit.