Fix VecGymNE so that it can work with adaptive popsize
The problem class `VecGymNE` used to fail when a
distribution-based search algorithm (such as
`PGPE` or `CEM`) used to fail with adaptive
population size (enabled with the setting
`num_interactions=...`) in non-distributed mode
(i.e. with `distributed` set as False).
The reason for this was that `VecGymNE` was not
reporting the number interactions and the number
of episodes it encountered in its status
dictionary. This commit fixes this mentioned
issue, allowing `VecGymNE` to work with
adaptive population size under non-distributed
mode.