Fix the "center" of CMAES to be 1-dimensional
Addressed issue:
https://github.com/nnaisense/evotorch/issues/110

During the initialization of CMAES, the center
point of the search was generated not as a vector
of length `n`, but as a tensor of shape `(1, n)`.
Because of this, the reported "center" solution
in the status dictionary also ended up with an
unexpected leftmost dimension with size 1.

This fix introduces a `squeeze()` operation on
the initial center tensor, and also a shape
verification, ensuring that the center tensor is
1-dimensional and has a correct length.
With these changes, the reported "center" in the
status dictionary becomes 1-dimensional.