Add `@distribute` decorator for parallelizing functions
**Motivation.**
Although EvoTorch has had a functional API
for a while, it did not have easy-to-use
parallelization capabilities for distributing
computations across multiple devices.
When writing an evolutionary search code using
the functional API, the only way to evaluate
solutions in parallel across multiple devices
was to combine the functional-paradigm code
with the object-oriented-paradigm code
(by instantiating a `Problem` object with
multiple actors and then transforming that
`Problem` object to an functional evaluator
using the method `make_callable_evaluator`.
This approach was both limiting (i.e. only
parallelized solution evaluation in mind)
and was cumbersome (forcing the programmer to
mix the usages of two APIs).

**The newly introduced feature.**
This commit introduces a general-purpose
decorator `evotorch.decorators.distribute`
which can take a function and transform it into
its parallelized counterpart. Like a Problem
object, the `distribute` decorator can be
configured in terms of number of actors
(`num_actors`) and number of GPUs visible to
each actor (`num_gpus_per_actor`).
Alternatively, an explicit list of devices can
be given (e.g. `devices=["cuda:0", "cuda:1"]`).

**How does it work?**
Upon being called for the first time, the
parallelized function will create ray actors.
When it receives its arguments, the parallelized
function will split those arguments into chunks,
and each chunk will be sent to an actor, the
wrapped function will be applied on the chunks
remotely in parallel on the devices associated
with those actors, and then finally the result
will be collected and combined.