Redesign `on_device`, `on_cuda`, `on_aux_device`
The decorators `on_device`, `on_cuda`, and
`on_aux_device` are re-designed so that they now
do functional transformation (i.e. they actually
do the moving of the tensors to the requested
device), instead of merely marking the decorated
functions. This is compatible with the newly
introduced `distribute` decorator, which also
does functional transformation instead of just
marking the decorated function.